backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.8.0
python-multipart==0.0.6
langchain==0.1.0
langchain-google-genai==0.0.6
google-generativeai==0.3.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
supabase==2.3.4
psycopg2-binary
sqlalchemy==2.0.23
requests==2.31.0
